LPS-63379 move to module and delete RulesEngineUtil
