+++ modules/portal-rules-engine/portal-rules-engine-api/src/main/java/com/liferay/portal/rules/engine/RulesEngine.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.bi.rules;
+package com.liferay.portal.rules.engine;
 
 import com.liferay.portal.kernel.messaging.proxy.MessagingProxy;
 import com.liferay.portal.kernel.messaging.proxy.ProxyMode;
