+++ modules/portal-rules-engine/portal-rules-engine-api/src/main/java/com/liferay/portal/rules/engine/RulesLanguage.java
@@ -12,29 +12,15 @@
  * details.
  */
 
-package com.liferay.portal.kernel.bi.rules;
-
-import java.io.Serializable;
+package com.liferay.portal.rules.engine;
 
 /**
  * @author Michael C. Han
  */
-public class Fact<T> implements Serializable {
+public enum RulesLanguage {
 
-	public Fact(String identifier, T object) {
-		_identifier = identifier;
-		_factObject = object;
-	}
-
-	public T getFactObject() {
-		return _factObject;
-	}
-
-	public String getIdentifier() {
-		return _identifier;
-	}
-
-	private final T _factObject;
-	private final String _identifier;
+	DROOLS_BRL, DROOLS_CHANGE_SET, DROOLS_DECISION_TABLE,
+	DROOLS_DOMAIN_SPECIFIC, DROOLS_DOMAIN_SPECIFIC_RULE, DROOLS_PKG,
+	DROOLS_RULE_FLOW, DROOLS_RULE_LANGUAGE, DROOLS_XML_LANGUAGE
 
 }
